MAINT: Rename folder structure for signal transformations
