+++ src/darsia/__init__.py
@@ -29,14 +29,14 @@ from darsia.corrections.shape.curvature import *
 from darsia.corrections.shape.drift import *
 from darsia.corrections.color.colorcorrection import *
 from darsia.corrections.color.experimentalcolorcorrection import *
-from darsia.transformations.signals.basemodel import *
-from darsia.transformations.signals.linearmodel import *
-from darsia.transformations.signals.thresholdmodel import *
-from darsia.transformations.signals.staticthresholdmodel import *
-from darsia.transformations.signals.dynamicthresholdmodel import *
-from darsia.transformations.signals.binarydataselector import *
-from darsia.transformations.reduction.signalreduction import *
-from darsia.transformations.reduction.monochromatic import *
+from darsia.signals.models.basemodel import *
+from darsia.signals.models.linearmodel import *
+from darsia.signals.models.thresholdmodel import *
+from darsia.signals.models.staticthresholdmodel import *
+from darsia.signals.models.dynamicthresholdmodel import *
+from darsia.signals.models.binarydataselector import *
+from darsia.signals.reduction.signalreduction import *
+from darsia.signals.reduction.monochromatic import *
 from darsia.regularization.tvd import *
 from darsia.regularization.binaryinpaint import *
 from darsia.analysis.translationanalysis import *
