+++ src/darsia/signals/models/basemodel.py
