+++ src/darsia/signals/models/binarydataselector.py
