+++ src/darsia/signals/models/linearmodel.py
