+++ src/darsia/signals/models/staticthresholdmodel.py
