+++ src/darsia/signals/models/thresholdmodel.py
