+++ src/darsia/signals/reduction/monochromatic.py
