+++ src/darsia/signals/reduction/signalreduction.py
